test: fix erroneously failing tests that were not having errors thrown
Some tests expecting an async function to reject and throw were failing with:

> Received function did not throw

The issue was intermittent when running all tests. Reproduction was reliable when running tests
serially in a single process (using Jest's `runInBand` option). The root cause is not fully
understood but matching against the shape of the error object is a workaround, instead of using
Jest's `toThrow` matcher. Some issues in Jest's issue tracker that may be related include:
jestjs/jest#13158 and jestjs/jest#12024.