fix(onboarding): fix keyboard covering footer button on onboarding screens
The change introducing the onboarding progress indicator removed the navigator header in favour
of using the navigator layout instead. This meant the header height passed to the vertical offset
prop on the `KeyboardAvoidingView` is always `0`, resulting in not enough space being left for the
keyboard.

To resolve, the keyboard inset reported by Unistyles can be used instead. There are improvements in
the ergonomics as the behaviour and offset props are not required. However, there is an issue where
the styles referencing the inset do not update unless the style sheet is defined in the same file
as the component – see the comment. When this issue is resolved in Unistyles, the style sheet can
be moved back to the dedicated style file.