refactor(devtools): allow opening of storybook via expo dev menu
Improvement over the current setup as it was only possible to have either the app or Storybook open
at one time. To switch between them, meant restarting the JS bundle server with the configuration
for one or the other. Now, Storybook is accessible via the Expo Dev Client menu.

Patch required on `@gorhom/bottom-sheet` to fix an issue with the bottom sheet throwing an error.
See gorhom/react-native-bottom-sheet#2155.