fix(onboarding): fix header jankily disappearing when opening onboarding modal screens
fa850e5 introduced a change which hid the custom navigator header when opening onboarding modal
screens as a fix for empty space being intermittently left under modal screens. That bug is no
longer reproducible (for an undetermined reason). As a result, that change can be removed as it
also causes some jank as the header disappears when an onboarding modal screen is opened.