build: disable support for web
Web is currently not supported however there was originally no harm from modifying the default
config from Expo. However, after the addition of the React Navigation dev tools plugin, it was
causing the following bundling error:

> Web Bundling failed x ms index.js (x modules)
>
> Metro error: Importing native-only module "react-native/Libraries/Utilities/codegenNativeCommands" on web from: node_modules/react-native-pager-view/lib/module/specs/PagerViewNativeComponent.js

The root cause for this is undetermined. Disabling support for web removes the error, deferring
investigation of this issue until support for web is required.