refactor(collection): move collection ui components from index files to files named after component
Index files improve file imports by allowing for more concise syntax as the component file name can
be omitted from the import path. However they make navigating difficult as there are lots of files
with the same name. Moving components to files named after the component resolves this. Index files
are kept but just export the component which keeps the benefits of more concise import paths. The
risk of increasing the time to bundle will be reduced by minimising exports from the index files
and the amount of branches in module graphs.