refactor(review): prepare next review query for returning multiple reviewables
Before this change, the next review query returns 1 review(able) at a time. This has been generally
sufficient. However, in the future, it will be possible to review things that are not due. This
allows for a situation where a reviewable gets returned as the next reviewable immediately after it
has been reviewed if the due date of other reviewables are further in the future. For example:

1. There are two reviewables with the following due _timestamps_:
   - `A`, due: `10`
   - `B`, due: `100`
2. `A` is reviewed and gets a new due timestamp of `20`. It is still older that `B` so gets returned
   as the next due reviewable.

Reviewing the same thing consecutively is undesirable. Querying for a fixed amount of reviewables to
be reviewed in a session will resolve this. In the above situation, `B` will be next thing to review
after `A` and the review session will complete after `B` has been reviewed.

This commit prepares for this change by:

- Updating the query function name to signify it can return multiple review(able)s.
- Updating the return type of the query function to return an array (while still only returning 1
  item).

It is split from actually implementing returning multiple items to simplify implementation and
testing.