refactor(note): group note fields in sql when getting notes
Improves performance of getting notes (by as much as 50% according to benchmarks) as getting a
note's fields are made in the same SQL statement as getting the note instead of individually
querying for fields for each note. There is a drop in performance when searching for a single or
non-existent note. However, it is worth it since this query will usually return multiple notes.