refactor(database): update how check constraints are defined
`drizzle-orm@0.36.0` deprecated the signature where the third parameter of the SQLite table function
is a function which returns an object (see https://github.com/drizzle-team/drizzle-orm/releases/tag/0.36.0).
The supported signature is a function which returns an array. The types, however, don't seem to allow
such a signature so upgrading `drizzle-orm` is necessary.

At the time of writing, `0.42.0` is the latest however an upgrade beyond `0.40.x` cannot take place
as `0.41.0` (see https://github.com/drizzle-team/drizzle-orm/releases/tag/0.41.0) introduces a
change in behaviour to blob columns which causes test failures.