perf(review): add benchmark for creating a review
Measures how long it takes to create a review for a reviewable depending on how many times it has
been reviewed in the past. The actual time for insertion should be constant however the operation
queries for things such as the latest snapshot for the reviewable, which is affected by how many
reviews have been completed in the past.