docs(patch): add reason for patching `@gorhom/bottom-sheet` in the patch file name and contents
Adding the reason in the patch file name allows for splitting patches into multiple files so that
each patch solves one specific issue. This improves the understanding of what a patch does and
improves maintainability as a specific patch can be deleted without interfering with other patches
should the source resolve an issue.