refactor(ui): update size and spacing tokens
Spacing tokens have been updated to use descriptions as their keys instead of an index on a scale,
making them easier to understand. Spacing values reference a point on a scale, to ensure arbitrary
values are not used. Various styles using values outside of the spacing system have been updated to
use a value within the system (some of these updates may be temporary as designs are not final).