build(commit): stop duplicate test runs
`lint-staged` was configured to run tests in separate tasks. One task was run when JavaScript files
changed and the other when TypeScript files changed. This was a problem when both JavaScript and
TypeScript files changed as both tasks run concurrently. This made tests run very slowly and prone
to failing due to timing out. The config has been updated to run tests in its own task when there
are changes to either JavaScript or TypeScript files.

The test command has also been updated to use the `findRelatedTests` option instead of `onlyChanged`
as it is more precise in running only the necessary tests. `passWithNoTests` has also been specified
to handle the case that there are no related tests.