refactor(database): upgrade `drizzle-kit` to `0.31.4` and `drizzle-orm` to `0.44.3`
`drizzle-orm@0.41.0` updated the `blob` type which necessitated adding a custom blob type. The
built-in `blob` from Drizzle was updated to always return a buffer while SQLite stores data in
the blob column exactly as input. We want our blob column to be able to store strings and blobs
and return them as is. This custom type enables this behaviour.

See:
- https://github.com/drizzle-team/drizzle-orm/releases/tag/0.41.0
- https://www.sqlite.org/datatype3.html