ci(test): add workflow to test source code
`.tool-versions` has been updated to work with the `setup-node` action. The empty line comment,
which served as a margin between two paragraphs, has been removed because the regular expression
that the `setup-node` action uses to find the Node version kept matching the `#` character. See:
https://github.com/actions/setup-node/blob/7e24a656e1c7a0d6f3eaef8d8e84ae379a5b035b/src/util.ts#L59.
The Node version has been set to `22` (the current LTS), as the `setup-node` action was not able to
understand `lts` – it expects `lts/*` however `mise` didn't seem to understand that.

The tolerance for the snapshot due date in the tests for the create review operation has been
increased to reduce flakiness.