test(e2e): replace `>` character in end-to-end test names with comma
The `>` character was being used a separator in test names to help identify groups of related tests.
However, it was preventing test debug artifacts from being saved during test failures in CI because
Maestro includes the name of the test in various files and GitHub Actions does not allow the use of
the `>` character in file names. The run fails with the following error:

> The path for one of the files in artifact is not valid [...]. Contains the following character:
>
> Greater than >
>
> Invalid characters include:  Double quote ", Colon :, Less than <, Greater than >, Vertical bar |,
> Asterisk *, Question mark ?, Carriage return \r, Line feed \n
>
> The following characters are not allowed in files that are uploaded due to limitations with
> certain file systems such as NTFS. To maintain file system agnostic behavior, these characters are
> intentionally not allowed to prevent potential problems with downloads on different file systems.

Replacing the `>` character with `,` resolves this while allowing for test names that are easily
read by humans and for groups of tests to be easily identified.