ci(test): add job to test ios app
End-to-end tests on iOS are now being run in CI. They are currently experimental and only advisory.
Failing job runs may need to be re-run or verified by running tests locally. This is due to the
flakiness of tests (for which the reason is undetermined, but may be due to limited resources on the
CI machines). To mitigate this flakiness, some notable things have been done:

- All end-to-end tests have been updated to automatically retry on failure, up to 3 attempts.
- End-to-end tests are split into shards to reduce the time of re-runs due to flakiness.
- The CI step that runs the tests can automatically retry after the shard has been completed, up to
  3 attempts.

The end-to-end test runner, Maestro, has been pinned at version `1.39.13` due to issues opening deep
links. See https://github.com/mobile-dev-inc/Maestro/issues/2610.