test(e2e): fix `clearState` variable defined in sub-flow overriding variable defined in caller
For an undetermined reason, the value of the `clearState` variable defined in the launch app
sub-flow (which serves as a default) was overriding the value defined by the flow calling the
sub-flow when running Android tests in CI. This was affecting the onboarding from app index
when not started flow which sets `clearState` to `true` but the default (`false`) was overriding
it. The reason is unclear because:

- The flow had been performing as expected in CI
- Maestro version has not changed
- There are no other changes that clearly affect this behaviour

Removing the default value, so that it is `undefined`, should stop `true` values being overridden.
It is just slightly less clear where the `clearState` value comes from.