build(android): enable gradle configuration cache
Enabling the [Gradle configuration cache][1] reduces build times as Gradle does not need to perform
the configuration phase from scratch. In CI, it has been observed to save approximately 10 minutes.
Some dependencies required patching to resolve [configuration cache violations][2].

The method of restoring and saving the Android NDK from the cache has been updated as enabling the
Gradle configuration cache broke the previous method.

To save some time on Android NDK install time, a dependency, `react-native-unistyles` was patched to
inherit the root project's Android NDK version as it did not have one specified so was defaulting to
the [default version][3] determined by the Android Gradle Plugin, which was not installed on GitHub
runners.

[1]: https://docs.gradle.org/8.13/userguide/configuration_cache.html
[2]: https://docs.gradle.org/8.13/userguide/configuration_cache.html#config_cache:requirements
[3]: https://developer.android.com/build/releases/gradle-plugin#compatibility