ci(build): increase ccache max size to 2 gb
After the ccache for iOS builds was evicted from the GitHub Actions cache, build times were
increasing. This was because the previous cache entry was created when the `max-size` was
increased during experimentation and was removed after experimentation. It should have been
set according to the size reported by `ccache --show-stats` – which this commit does.