ci: remove disk clean up from android jobs
Cleaning up the disk for Android jobs was added to stop jobs failing due to out of disk space errors
which presented themselves like this:

> Unhandled exception. System.IO.IOException: No space left on device

While it solved the issue, it added approximately 2 minutes to job run times – the most amount of
time after the build step or test step. With more experimentation, it seems it can now be removed.
Why it can now be removed is undetermined but I suspect updating the build to only build for the
active architecture reduces the amount of disk space required.