style(yaml): use default yamllint rules
Default yamllint rules are stricter. `document-start` has been disabled because it is optional and
not typically used. `truthy` has been disabled because GitHub Actions workflows uses `on` as a key,
which can be interpreted as `true` by older YAML parsers however this is not a problem with GitHub
Actions. The permitted has been increased to `120` because the default `80` is too strict for some
YAML documents which contain a bit of prose (e.g. GitHub Actions input descriptions).

YAML documents have been updated to conform with the new rule configuration.