ci: configure check workflow to run on push events to only the main branch
The check workflow was configured to run on pushes to every branch. For pull requests it would run
on both push and pull request events. The intention was validating changes worked, by building and
testing with the state of the branch on push; and validating changes could be merged successfully,
by building and testing with the state of the branch merged into the main branch. The push event
handles the former, while the pull request event handles the latter.

The problem is running all these jobs hits the [concurrency limit][1] of GitHub Actions. This would
cause queues in job runs which could result in total (queue + run) times 2-5 times greater than the
nominal of approximately 30 minutes.

To solve this, workflow runs on pushes to branches other than main are being removed. Running checks
with the state of the branch merged is more important. Additionally, checks with the isolated state
of the branch can be done manually (either locally or triggering in CI manually). This means, for
branches without a pull request, checks are only run locally, manually triggered in CI, or once a
pull request is opened. For branches with pull requests, the checks will still run on synchronize.

[1]: https://docs.github.com/en/actions/reference/limits#job-concurrency-limits-for-github-hosted-runners