ci(deps): limit dependabot to 1 pull request per day
Dependabot was configured to run monthly with the default limit for maximum number of pull
requests, 5. The problem was opening 5 pull requests would create a queue of checks to build
and test the proposed changes. Once one pull request was merged, all the other ones would then
be rebased which again triggers checks. This resulted in queues taking hours long.

Making Dependabot checks daily but limiting the amount of open pull requests to 1 should spread
the load, reducing the amount of CI runs and rebases which trigger CI runs.