ci(deps): make dependabot ignore major and minor version updates from react and react-test-renderer
React's (and therefore React Test Renderer's) major and minor versions should be compatible with
React Native. For example: React Native >=0.78 is compatible with React 19.0 and React Native >=0.80
is compatible with React 19.1. As a result React and React Test Renderer should have their major and
minor versions ignored by Dependabot and be grouped with upgrades to React Native.