ci(build): fix android build failing due to missing `build` directory
The native dependency linking step was failing due to the `android/build` directory not existing.
Unsure why it no longer exists given the expectation was that it would be created when the Expo
prebuild command is run and no dependencies have changed meaning that behaviour is not expected to
change. Creating the missing directories, if they do not exist, fixes this and also makes the
process more resilient.